user_service/requirements.txt
databases==0.9.0
fastapi==0.115.4
pydantic==2.9.2
SQLAlchemy==2.0.36
uvicorn==0.32.0
python-dotenv~=1.0.1
aiosqlite==0.20.0
pika~=1.3.2
